incorporate site into header fix database
Visit numbers are not sufficient to identify data, since the
different sites (generally just Subaru, LAM and simulator, but
could be others in theory) use separate visit counters. Now
visit=54 for Subaru requires header modification, but that
clashes with visit=54 from the simulator (weekly dataset)
which breaks with that modification.

Added site into the header fix database, so we can distinguish
between these. Subaru fixes (which are the main ones we care
about) are written to PFS-<visit>.yaml, while the others are
written to PFS-<site>-<visit>.yaml.