Black dot optimization algorithm
This is a class containing functions that can optimize black dot positions. The algorithm takes mcs data, and original dot positions, as an input and gives `penalty` score for modified positions of the dots. The penalty is calculated a number of points which are seen in the data, although they should be covered given the dot position + number of points which are not seen in the data, although they should not be covered given the dot position.