Fix underlying problems with initial distortion fits
When some of the "ring" fiducials the initial fitter can come up
with a solution that mis-asigns fibres. Fix this by adding a prior
on dtheta.

N.b. it is the users responsibility to use this!  E.g.
    alphaRot = 1
    pt = transformUtils.fromCameraName('rmod', altitude=90, insrot=insrot, nsigma=0, alphaRot=alphaRot)

    outerRingIds = [29, 30, 31, 61, 62, 64, 93, 94, 95, 96]
    fidsOuterRing = fids[fids.fiducialId.isin(outerRingIds)]

    pt.updateTransform(mcsData, fidsOuterRing, matchRadius=8.0, nMatchMin=0.1)
    pt.nsigma = nsigma
    pt.alphaRot = 0

    for i in range(2):
        fid, dmin = pt.updateTransform(mcsData, fids, matchRadius=4.2, nMatchMin=0.1)

Also exposed nsigma and alphaRot to the pfiTransform ctor, and improved the plotting a little