Support doHex=True, makeSingleHexDither=True
Also fix a bug which failed to write the design name to variant=0

This actually only supports a 7-point dither, so I expect that
you will call this with variant=0..6

N.b. this is _not_ the code to make sets of e.g. 7 cobras trace
out a full 7-point hex dither.