Move the location of importing the eups package
fiberids.py in python/pfs/utils/ in pfs_uitls imports eups at the
top section, but it is only used to look up the PFS_UTILS parameter
using eups.productDir('PFS_UTILS'). As the constructor of FiberIds
allows to be called with a path to the fiberid data directory,
it would be nice if the import line is moved right above the line
using eups to minimize the dependency to eups for Python environments
not using it. This commit enables it.