Add a path to fiberid data to makePfsDesign()
I'm trying to modify the package so that it can be used outside of
LSST stack (i.e., EUPS-managed environemnts). It seems only FiberIds()
depends on EUPS in pfs_utils. So, I added `fiberidsPath` to
`makePfsDesign` to make it possible to use without EUPS by providing
the path to `pfs_utils/data/fiberids` directory.

It seems that `FiberIds()` is hard-coded without any keywords in
`blackDotsScript.py`, but for my current purpose, it is not essential.